Refactor registration components
